refactored RotateNMS to layers.rotate_nms.py
